+++ src/main/kotlin/io/acari/doki/icon/patcher/PatcherModels.kt
@@ -1,0 +1,17 @@
+package io.acari.doki.icon.patcher
+
+import java.util.*
+
+class PathPatcherDefinition(
+  val name: String,
+  val pathToRemove: String,
+  val pathToAppend: String
+)
+
+class PathPatchersDefinition(
+  val glyphs: List<PathPatcherDefinition>
+)
+
+class MaterialPathPatchers(
+  val glyphs: List<MaterialPathPatcher> = Collections.emptyList()
+)