+++ src/main/kotlin/io/acari/doki/icon/provider/IconProvider.kt
@@ -1,0 +1,7 @@
+package io.acari.doki.icon.provider
+
+import javax.swing.Icon
+
+interface IconProvider {
+  fun getIcon(virtualFileInfo: VirtualFileInfo): Icon?
+}