+++ src/main/kotlin/io/acari/doki/notification/UpdateNotification.kt
@@ -6,14 +6,17 @@ import com.intellij.openapi.project.Project
 val UPDATE_MESSAGE: String = """
       What's New?<br>
       <ul>
+      <li>Doki Themed Material Icons, see settings to enable them.</li>
       <li>Brought back configurable theme name in Status Bar.</li>
+      <li>Updated Light Natsuki's color scheme.</li>
+      <li>Themed the SSV Normandy Progress Bar Icon.</li>
       </ul>
       <br>Please see the <a href="https://github.com/cyclic-reference/ddlc-jetbrains-theme/blob/master/changelog/CHANGELOG.md">Changelog</a> for more details.
       <br>
       Thanks again for downloading <b>The Doki Theme</b>! •‿•<br>
 """.trimIndent()
 
-const val CURRENT_VERSION = "5.0.3"
+const val CURRENT_VERSION = "5.1.0"
 
 object UpdateNotification {